third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/anonymous-iframe/local-storage.tentative.https.window-expected.txt

This is a testharness.js-based test.
PASS Setup
FAIL same_origin anonymous iframe can't access the localStorage assert_equals: expected "" but got "same_origin"
FAIL cross_origin anonymous iframe can't access the localStorage assert_equals: expected "" but got "cross_origin"
Harness: the test ran to completion.
